Removed unnecessary testing_helpers
