+++ Makefile
@@ -6,7 +6,7 @@ run:
 test:
 	cd goPageMaker && go test -cover
 
-testCoverageHTML:
+coverage:
 	cd goPageMaker && go test -cover && go tool cover -html=cover.out
