+++ goPageMaker/helpers_test.go
@@ -1,6 +1,9 @@
 package main
 
 import (
+	"errors"
+	"fmt"
+	"reflect"
 	"strconv"
 	"testing"
 )
@@ -41,4 +44,43 @@ func TestConvertToInteger(t *testing.T) {
 
 		assertError(t, got, want)
 	})
+}
+
+func assertEquals(t testing.TB, expected, result fmt.Stringer) {
+	t.Helper()
+	if reflect.DeepEqual(expected, result) {
+		return
+	}
+
+	t.Log(expected.String(), result.String())
+
+	t.Errorf("Variables are not equal, \nexpected: %s \nresult: %s", expected, result)
+}
+
+func assertEqualsInt(t testing.TB, expected, result int) {
+	t.Helper()
+
+	if expected == result {
+		return
+	}
+
+	t.Log(expected, result)
+
+	t.Errorf("Integers are not equal. \nexpected: %d \nresult: %d", expected, result)
+
+}
+
+func assertError(t testing.TB, got, want error) {
+	t.Helper()
+
+	// log.Fatalf("\nexpected %q \ngot %q", want, got)
+	if !errors.Is(got, want) {
+		t.Fatalf("got error %q want %q", got, want)
+	}
+
+}
+
+func assertNoError(t testing.TB, err error) {
+	t.Helper()
+	assertError(t, err, nil)
 }
