--- goPageMaker/testing_helpers.go
@@ -1,47 +1,0 @@
-package main
-
-import (
-	"errors"
-	"fmt"
-	"reflect"
-	"testing"
-)
-
-func assertEquals(t testing.TB, expected, result fmt.Stringer) {
-	t.Helper()
-	if reflect.DeepEqual(expected, result) {
-		return
-	}
-
-	t.Log(expected.String(), result.String())
-
-	t.Errorf("Variables are not equal, \nexpected: %s \nresult: %s", expected, result)
-}
-
-func assertEqualsInt(t testing.TB, expected, result int) {
-	t.Helper()
-
-	if expected == result {
-		return
-	}
-
-	t.Log(expected, result)
-
-	t.Errorf("Integers are not equal. \nexpected: %d \nresult: %d", expected, result)
-
-}
-
-func assertError(t testing.TB, got, want error) {
-	t.Helper()
-
-	// log.Fatalf("\nexpected %q \ngot %q", want, got)
-	if !errors.Is(got, want) {
-		t.Fatalf("got error %q want %q", got, want)
-	}
-
-}
-
-func assertNoError(t testing.TB, err error) {
-	t.Helper()
-	assertError(t, err, nil)
-}
